Piers' bugfixing and animation
I've fixed the issue of fireflies giving the wrong amount of score, it was because of the order the Player's collision detector was looking for things in. I've also increased the score the player needs to win from 5 to 10, and made fireflies worth twice as many points as flies are.

I also added code into all four actors to create walking/flying/hopping animations, and tweaked the way MyWorld now creates more flies over time (making the Spawner class redundant).